backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
olca-ipc
pydantic
python-dotenv
sqlalchemy
pytest
